Move CCoins-related logic to coins.{cpp.h}
